requirements.txt
fastapi
python-dotenv
uvicorn
kokoro
soundfile
pydantic_ai
pydantic
